Use a dict instead of function
