+++ scripting/src/test/tweakflow/exception.tf
@@ -4,9 +4,9 @@
 library echopraxia {
 
   doc 'Evaluates if exception matches'
-  function evaluate: (string level, function ctx) ->
+  function evaluate: (string level, dict ctx) ->
     let {
-      find_string: ctx("find_string");
+      find_string: ctx[:find_string];
     }
     find_string("$.exception.message") == "testing";
 }
